Ported over GCC code to Windows
Using C (original 1989 standard) instead of C++.
Modified the code to make it compile and run on Windows.
Appears to work successfully at this point.
The issue of a long timeout on unsuccessfull connection
still remains, which only happens with Windows sockets.
Also added a visual studio makefile (execute with nmake)
to build the static library and executable on the
command line.