Reorganized makefiles and build directories
One common makefile for Windows and UNIX now,
for two split into two, but the only difference
is the OS variable at the top. Eventually this
will be passed in by the script calling "make".
The two files will be combined into one at the root
of the directory. It will know to use build-windows
or build-unix based on the OS variable.